twamp-server.service: Support 4096 control procs
DANOS now allows the maximum-connections per TWAMP server to be
configured up to 4096. This requires an increase in the TasksMax
value to allow for 4096 control processes, 4096 reflector processes
and the main process.

The file descriptor limit must also be increased to allow for this
number of connections. LimitNOFile is therefore doubled which gives
lots of headroom.

VRVDR-39564