twping: Allow more immediate start time
Currently twping estimates a session start time based on an estimated
round trip time for the remaining control messages plus 1 second.
This can be overidden by using the start delay (-z) option, however
this will only take effect if the specified delay is greater than
this estimate.

Since the TWAMP client only starts sending test packets after it has
received the Start-Ack message from the server we know that both
sides are ready and willing to start the session as soon as the
client receives this message. Therefore we can allow twping to be
configured to use a more immediate start time, such that the session
is likely to start as soon as the Start-Ack is received.

This is done by relaxing the start delay restriction for twping. Now,
if the provided start delay is less than the estimate the start time
is re-calculated using the estimated round trip time for the
remaining control messages plus the provided start delay value (rather
than the previous hardcoded 1 second).

VRVDR-49423