schedule: Add support for slot flags
Today the send scheduling behaviour for TWAMP sessions mirrors that
for OWAMP sessions. OWAMP scheduling is defined by the protocol since
both ends need to be aware of the schedule. However, for TWAMP the
reflector does not need to know anything about the send schedule of
the sender, therefore we are free to implement alternative scheduling
behaviour.

Therefore this commit adds support for "slot flags" which can be used
to alter the scheduling behaviour for TWAMP sessions.

No new behaviour/flags are defined in this commit, we are simply
adding support to the library.

To support this OWPSlot gains a flags field to allow flags to be set
on a per-slot basis. When writing the schedule slots to the data file
we steal the unused top 4 bits of the slot type field to encode up to
4 flags for each slot. This is done by _OWPEncodeNonStandardSlot()
and _OWPDecodeNonStandardSlot() which are used when reading/writing
the data file.

VRVDR-49424