schedule: "cycle to" scheduling for TWAMP
Today the send scheduling behaviour for TWAMP sessions mirrors that
for OWAMP sessions. The schedule is determined by defining a series
of slots, with each slot representing a delay value. When sending a
test packet we look at the current slot and wait for the delay period
prior to sending the packet. Before sending the next test packet we
move to the next slot and repeat the process. Once we reach the final
slot we circle back to the first slot for the next packet, and this
continues for the duration of the test.

This commit adds support for "cycle to" schedules for TWAMP sessions.
A "cycle to" schedule is one where we do not cycle back to the first
slot once we reach the final slot, we instead cycle back to a later
slot which can be specified by the setting of a flag on the slot.

To enable a "cycle to" schedule the TWP_SLOT_FLAG_CYCLE_TO flag
should be set on the OWPSlot representing the slot which should be
cycled to each time the last slot is reached.

This flag must not be set for OWAMP schedules since the scheduling
behaviour is part of the protocol, defined by the RFC.

VRVDR-49424