twping: Add support for "cycle to" schedules
A "cycle to" schedule is one where we do not cycle to the first slot
once we reach the final slot. Instead we cycle to a slot specified by
the user in the schedule definition.

A "cycle to" schedule is specified by delimiting a slot with a
semicolon rather than comma. The slots before the semicolon are used
only once on the first pass through the schedule. The remaining slots
are used cyclically throughout the remainder of the test.

For example, the schedule "-i 0f,1f;2f" will send the first packet as
soon as the session starts, the second packet after 1 second, and all
remaining packets every 2 seconds thereafter.

These schedules are useful when we want to specify no delay for the
first packet in a test, or perhaps to specify a slower packet rate
at the start of the test which then ramps up as the test progresses.

owping does not accept a non-cyclic schedule and will exit with an
error if one is provided.

VRVDR-49424