twping: implement test end delay
twping currently exposes a test end delay option (-E) but it does
not currently have any effect. This commit adds support for the
end delay, with the following semantics:

When configured, the end delay determines the period to wait before
sending the Stop-Sessions message once all test packets, which have
not already been marked as lost, have been replied to. The delay will
not come into effect until we have attempted to send all test packets
for the session.

When the end delay is not configured, or prior to this commit, we
would wait for the loss timeout period to expire before terminating
the session.

This allows twping to be configured to end successful tests more
quickly when we do not care for duplicate statistics, while still
having a reasonably long loss threshold value.

In other words, when an end delay is specified it takes over from
the loss threshold timeout when all test packets have been received.

VRVDR-49425