twamp: allow vyatta_endpoint_session_end() to be called multiple times
The dataplane session needs to be deleted following a control-plane
failure, e.g. a request for too many sessions. Call
vyatta_endpoint_session_end() whenever the sessions is terminated
(_OWPEndpointStop()).

However that means deletion can occur twice, once when the test
completes and again as part of session cleanup
(_OWPEndpointStop()). Use the offload indicator to determine if the
session needs to be removed from the dataplane.

Fixes: VRVDR-55606